refactor: 수익 계산 로직 이동 및 Scoreboard 클래스 생성
