+++ src/main/kotlin/blackjack/domain/model/participant/Player.kt
@@ -2,10 +2,10 @@ package blackjack.domain.model.participant
 
 import blackjack.domain.model.Action
 import blackjack.domain.model.Bet
-import blackjack.domain.model.GameResult
 import blackjack.domain.model.card.Card
 import blackjack.domain.model.card.Deck
 import blackjack.domain.model.card.Hand
+import blackjack.domain.model.result.GameResult
 
 class Player(
     name: String,
