+++ src/test/kotlin/blackjack/domain/model/DealerTest.kt
@@ -5,6 +5,7 @@ import blackjack.domain.model.card.Rank
 import blackjack.domain.model.card.Suit
 import blackjack.domain.model.participant.Dealer
 import blackjack.domain.model.participant.Player
+import blackjack.domain.model.result.Scoreboard
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
